+++ CHANGELOG.md
@@ -1,4 +1,16 @@
-## [1.10.0](https://github.com/GodIsI/homebridge-ups-monitor/compare/v1.9.1-beta.224...v1.10.0) (2026-06-01)
+## [1.10.7](https://github.com/GodIsI/homebridge-ups-monitor/compare/v1.10.6-beta.252...v1.10.7) (2026-06-04)
+
+### Features
+
+* **export:** add battery % and runtime to 30-day CSV ([0f56fff](https://github.com/GodIsI/homebridge-ups-monitor/commit/0f56fff0e88826ee3c8c9db0c01fa341ff336888))
+
+### Bug Fixes
+
+* **dashboard:** draw real voltage instead of smoothing it away ([1249469](https://github.com/GodIsI/homebridge-ups-monitor/commit/1249469f9bf1f09e1403b5e4334d334b2795a987))
+* **lifecycle:** clean up poll timers and dashboard server on shutdown ([a3d0568](https://github.com/GodIsI/homebridge-ups-monitor/commit/a3d0568805c1e42e648936a05fb1d77e09a87ad7))
+* **poll:** validate interval and prevent overlapping polls ([ef487b7](https://github.com/GodIsI/homebridge-ups-monitor/commit/ef487b79443e9fca362a6995639f192c815f90c6))
+
+## [1.10.6-beta.252](https://github.com/GodIsI/homebridge-ups-monitor/compare/v1.10.5-beta.248...v1.10.6-beta.252) (2026-06-01)
 
 ### Bug Fixes